convert to int

+++ main/generator.py
@@ -840,7 +840,7 @@ def generate_assertion(assertion_temp, klee_dir):
 
 
 def generate_extended_patch_formula(patch_formula, path_condition):
-    angelic_count = len(re.findall("angelic!bool!(.+?)!0", path_condition.serialize())) / 4
+    angelic_count = int(len(re.findall("angelic!bool!(.+?)!0", path_condition.serialize())) / 4)
     if angelic_count == 0:
         return patch_formula
     model_path = generate_model(path_condition)
